% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\docType{data}
\name{.hallmark_cache}
\alias{.hallmark_cache}
\title{Prepare Seurat data for spatial selector}
\format{
An object of class \code{environment} of length 2.
}
\usage{
.hallmark_cache
}
\arguments{
\item{seurat_input}{Seurat object or data list}

\item{sample_name}{Character label for dataset}

\item{show_image}{Logical; whether to show image}
}
\value{
A list containing processed Seurat object and metadata
}
\description{
Prepare Seurat data for spatial selector
}
\keyword{internal}
