% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_SpatialROI}
\alias{run_SpatialROI}
\title{Launch SpatialROI}
\usage{
run_SpatialROI(data_path = NULL, ...)
}
\arguments{
\item{data_path}{Optional path to a user data file (.rds) or "demo" to load example data.}

\item{...}{Additional arguments passed to \code{shiny::runApp()}.}
}
\description{
Launch the Shiny interface for SpatialROI. Users can either:
\itemize{
\item Call \code{SpatialROI::run_SpatialROI()} to open the app and upload their own data interactively, or
\item Call \code{SpatialROI::run_SpatialROI("path/to/data.rds")} or \code{SpatialROI::run_SpatialROI("demo")} to launch the app preloaded with a dataset.
}
}
\examples{
\dontrun{
SpatialROI::run_SpatialROI()
SpatialROI::run_SpatialROI("demo")
SpatialROI::run_SpatialROI("~/Downloads/mydata.rds")
}

}
